Initial work on NeoForge port
Currently not working.

The main idea is to transform and load `org.lwjgl.glfw.GLFW`, without mixin and before the class is used. The class is transformed using the awesome ClassTransform library.

We are currently able to inject into `glfwInit()` as a proof of the viability of this approach, but these issues remain to be resolved:

- Mod classes are not visible to the transformed GLFW class, which resides on the `BOOT` layer
- Ensure correct handling of the early display window (we might be able to start working only after the window is handed to Minecraft)
- The mod needs to be correctly structured using JiJ