Add Forge support
Use `Instrumentation` to redefine GLFW class which is already loaded when `ITransformerService`s are run. The Reflect library is used to obtain an instance of `Instrumentation`, which surprisingly works even on JREs!
The tricky thing here is that methods cannot be added when redefining classes. Unfortunately, we had been using a lot of lambdas, which are also considered methods. Hence the weird workaround code in `LambdaHelper`.
Had to hardcode modid to make a gradle plugin happy.
TODO: Test with splash screen disabled. Test allocation rate.